requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
scikit-learn==1.7.2
scipy==1.16.2
pillow
threadpoolctl==3.6.0
Werkzeug==3.1.3
